// src/app/logistics/client/client.ts
export class Client {
  id: number;
  clientName: String;
  clientCode: String;
  pincode: String;
  contactPerson: String;
  mobile: String;
  email: String;
  active: number;
  // created fiels
  token: any;
  registeredAddress: any;
  awbSeriesPrefix: any;
  awbSeriesSequence: any;
  bankName: any;
  accountNo: any;
  ifscCode: any;
  beneficiry: any;
  city: string;
  state: string;
  country: string;
  latLong: String;
  panNumber: String;
  aadhaarNumber: String;
  gstNumber: String;
  accountManager: String;
  saleManager: String;
  docUploaded: number;
}